pyproject.toml
[project]
name = "poseidon-bot"
version = "0.1.0"
description = "Discord server monitoring bot"
requires-python = ">=3.13"
dependencies = [
  "discord.py>=2.4.0,<3.0.0",
  "psutil>=5.9.0,<7.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"
